+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Support WETH deposit and withdraw function in `HWeth`.
 - Support Curve deposit to gauges in `HCurveDao`.
 - Support Curve mint and mint_many CRV in `HCurveDao`.
+- Support Mooniswap add/remove liquidity function in `HMooniswap`.
 
 ## [0.8.0] - 2020-08-10